// src/components/arcade/sprite-atlas.generated.ts
export type BernardSpecialPose = "specialCharge" | "specialIce" | "specialCircle" | "specialOrb" | "specialScan" | "specialBreaking"
export type SpritePose = "idle" | "crouch" | "punchHigh" | "punchLow" | "kickHigh" | "kickLow" | "jump" | "ranged" | "down" | "walk" | "specialUp" | BernardSpecialPose